unify responsive layout in date filters panel panel owns width measurement; children receive isCompact/isWideScreen and compact styles cascade from the is-compact root, which stacks the controls vertically
