+++ projects/packages/premium-analytics/packages/ui/src/date-range-layout.ts
@@ -1,10 +1,4 @@
 /**
- * Threshold width (in pixels) below which preset controls collapse to a select.
- * Based on the container width, not the viewport.
- */
-export const MOBILE_CONTAINER_WIDTH_THRESHOLD = 480;
-
-/**
  * Threshold width (in pixels) for showing two months in the custom-range calendar.
  */
 export const WIDE_CALENDAR_CONTAINER_THRESHOLD = 560;
